% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trailR.R
\name{fingerprint}
\alias{fingerprint}
\title{---
title: "R functions for having an audit trail"
author: "Wilson, Bokov"
date: "03/21/2018"
---}
\usage{
fingerprint(filename = NA)
}
\value{
A list with elements that include...

   * `in.repo`: whether or not the current working directory is a git repo
   * `repo`: URL of the git repo, NA if `in.repo` is false
   * `branch`: currently checked-out branch
   * `branch.on.origin`: whether the currently checked-out branch has already
      been pushed to origin... because that's not a given and we don't want this
      scenario preventing reproduceability from taking us by surprise when we
      least expect it
   * `last.branch.commit`: the hash of whatever is currently checked out
   * `uncommitted.changes`: whether or not there are uncommited changes to
     any tracked files
   * `file.tracked`: if the optional `filename` argument is used, whether or not
     that file is being tracked by git.
   * `file`: the value of the optional `filename` argument.
   * `file.hash.or.last.commit`: if `file.tracked` then the hash for the last
     commit THAT INCLUDED CHANGES TO `file` (which is not necessarily the
     same thing as `last.branch.commit` if the file has not been changed during
     the most recent commit). Otherwise, it's the `md5sum()` of the file. If the
     file doesn't exist, this value is `NA`.
   * `timestamp`: when `fingerprint()` ran (doesn't tell you anything about the
     file or the repo themselves).
}
\description{
This is going to get called via getOption('trail',tinit())
The fields of trail are:
time = timestamp
type = type of information (file, rdata, info, gitstamp, seed, etc)
name = the name of that specific item, should be unique
value = Almost anything. Non atomic values should be assigned for example like this:
  `trail[nrow(trail),'value'] <- list(list(mtcars))`
  For items of type 'file' the path will be stored here, for 'seed', the random seed, for 'gitstamp'
  the current branch.
hash = For 'file' and 'rdata' the md5 checksum of the file, for 'gitstamp' the commit hash
trail = For 'rdata', if a trail object exists, dump it into this field. And it can have its own trail objects
  so we have an arbitrarily long "family tree" of ancestor objects!
}
\details{
Get the hash of a file, using the github one if it exists

\code{fingerprint()} is supposed to be a more flexible and robust way to get
reproducibility info about a file, and allows the same syntax to be used on
any file whether it is tracked or not, and even if the file doesn't actually
exist. In the future I plan to rewrite [gitstamp], [tload],
and \code{tread()} to all use `fingerprint()` internally but as of right
now they are independent.
}
