% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trailR.R
\name{gitstamp}
\alias{gitstamp}
\title{Return a git commit hash}
\usage{
gitstamp(production = T, branch = T)
}
\arguments{
\item{production}{If `TRUE` (default) then will throw an error if it finds
uncommitted changes. Otheriwse will return
`'TEST_OUTPUT_DO_NOT_USE'` as the git stamp.}

\item{branch}{If `TRUE` (default) also returns what branch is currently
checked out.}
}
\value{
character vector of branch and git stamp or just git stamp if
        `branch` is `FALSE`

Return a commit hash (for inclusion in reports for example) after first
making sure all changes are committed and pushed
}
\description{
Return a git commit hash
}
