% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trailR.R
\name{print.walktrail}
\alias{print.walktrail}
\title{Print function for `walktrail()`}
\usage{
\method{print}{walktrail}(xx, sub = c("sequence", "time", "type",
  "name_value", "hash"), ...)
}
\description{
Print function for `walktrail()`
}
