% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trailR.R
\name{tinit}
\alias{tinit}
\title{this creates a trail object if there isn't already one and either way
returns a trail object}
\usage{
tinit(trail = getOption("trail"), ...)
}
\description{
this creates a trail object if there isn't already one and either way
returns a trail object
}
