% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trailR.R
\name{tload}
\alias{tload}
\title{loading an rdata file and checking whether it has trail, to include in the
the current trail as a nested data.frame}
\usage{
tload(file, envir = parent.frame(), verbose = FALSE,
  trailobj = ".trail")
}
\description{
loading an rdata file and checking whether it has trail, to include in the
the current trail as a nested data.frame
}
