% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trailR.R
\name{tread}
\alias{tread}
\title{wrapper for most read functions, and records the file and its hash in trail.
Will eventually also check for accompanying flat-file trail files in JSON
format.}
\usage{
tread(file, readfun, ...)
}
\description{
wrapper for most read functions, and records the file and its hash in trail.
Will eventually also check for accompanying flat-file trail files in JSON
format.
}
