% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trailR.R
\name{tsave}
\alias{tsave}
\title{Wrapper for save(). Pulls trail out of options as a data.frame, saves the
data.frame along with whatever was originally going to be saved, and then
deletes it. Logs the save to itself before saving.}
\usage{
tsave(..., list = character(), envir = parent.frame(),
  trailobj = ".trail", verbose = TRUE)
}
\description{
Wrapper for save(). Pulls trail out of options as a data.frame, saves the
data.frame along with whatever was originally going to be saved, and then
deletes it. Logs the save to itself before saving.
}
