% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trailR.R
\name{tself}
\alias{tself}
\title{script registering itself... adds a gitstamp and its own name to trail}
\usage{
tself(scriptname = parent.frame(2)$ofile,
  production = getOption("gitstamp_prod", T))
}
\description{
script registering itself... adds a gitstamp and its own name to trail
}
