% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trailR.R
\name{tupdate}
\alias{tupdate}
\title{The function that updates the trail object with a new row of data}
\usage{
tupdate(type = NA, name = NA, value = NA, hash = NA,
  time = Sys.time(), trail = getOption("trail", tinit()),
  whichrecord = nrow(trail) + 1, parent.trail = NA)
}
\description{
The function that updates the trail object with a new row of data
}
